transform lookup - not working yet
